backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# AI Models
openai==1.10.0
anthropic==0.18.1
httpx==0.26.0

# Vector Database
chromadb==0.4.24

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
email-validator==2.1.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/ML
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
tiktoken==0.5.2
numpy==1.26.3

# Document Parsing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Monitoring and Logging
loguru==0.7.2
